speach/pyproject.toml
[project]
name = "speach"
version = "0.1.0"
description = "Voice recording app with transcription, image generation and feedback"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.37.4",
    "groq>=0.18.0",
    "openai>=1.65.2",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "streamlit>=1.42.2",
    "sounddevice>=0.4.6",
    "scipy>=1.12.0",
    "numpy>=1.26.2",
    "pillow>=10.2.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "pydub>=0.25.1",
    "pyaudio>=0.2.13",
    "streamlit-webrtc>=0.47.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build_backend"
